patch 0.8
Added support for best pick of two polygon points for each point in space.

There is stil la problem calculating the distances somehow, need to delve inside, works find on 4 point polygons though

coloring needs to be adjusted to chaos  and pattern aswell.

A json should be created that holds all data of painting for people that purchase