Smp-patch-0.2
Almost completed the center finder function in math_zone. all i need now is to make sure the radius is adequate and i can continue.

To do list:

Make sure the complete funnel of polygon creation is up and running.

Work on Numpy image distances and in-polygon coloring

Create a coloring class and a light fx class.