smp-patch-0.3
Center pinning works well, problem is that the covered edges system is still not fully fleshed out. what needs to be done is: figure a way to include an interval for t for each edge, instead of 0 to 1, it should start according to the radius that sits on top of it.