Patch 1.0 - For Review
Fixed confusion between relevant centers and radius dataframe that is in charge of the t_min_max with already existing circles; and rel_const which is in charge of the relevant triangles that the new center cannot manifest inside them.

also fixed some indexing issues when concating dataframes.

problem is: it still wont recognize that it is created inside a triangle (i.e 'below' an existing shape)